Add initial theming for styled-components
For the type definitions for styled-components (@types/styled-components), there is a major issue.
We cannot compile using their type defintion because the web definitions
conflict with react-native (mobile) definitions. This is a library
issue and cannot be fixed unless we modify the compiler options in
tsconfig.json or forcibly remove the react-native definitions.

We decide to foribly remove the definitions using `yarn autoclean
--force`, while providing the .yarnclean file.

$ cat .yarnclean
@types/react-native

Thread discussing the issue: https://github.com/DefinitelyTyped/DefinitelyTyped/issues/33311