Connect async action creators to dispatch
Connecting react-thunk with typescript project has been the most
horrible experience ever. Somehow this commit works, so use this
commit as a base to see why it worked. At the moment, we are using
the `any` type to call the dispatched action. We need to find a
more suited type later.