Protobuf Python Issue: Change directory structure
We need to do this because the protobuf tool doesn't yet compile the correct relative paths for python files.

Currently, using this hack https://github.com/grpc/grpc/issues/9575#issuecomment-293934506

Apparently, there is a fix, but it doesn't work for us since we need a central repo to handle all the protobufs.
https://github.com/protocolbuffers/protobuf/issues/1491#issuecomment-415505938

Most likely, we will switch to a hacky bash scrip with sed and awk to fix the generated stubs...