Cannot run globally when installing.
Running python setup.py install will package everything into an egg and put it in the right path. However, we are unable to run it because we get an import error from not being able to find libtorrent.

I've extracted the .egg and see that the libtorrent.pyd exist. The error also seems to be always looking for a libtorrent/__init__.py file when it shouldn't. I'm not sure why running the egg does that. This is will become a future task item to do if there are reasons to install ultimate-torrent as python module, globally or locally, for testing, developing, or any reason really .