Connect rpc from gateway to myanimelist GetUserAnimeList
Also, I have no fucking idea, but I was debugging this gateway for hours because it was decoding the http request and giving literally an empty body/proto message back so when it was sent to the gRPC service, the gRPC service would receieve literally a blank message with all fields zero'ed.

I start running the debugging and tracing everything, and then stop debugging, and now it magically works???? If this same issue occurs on a system for a new build, come back and look at this and try something, I don't know.